requirements-corex-extra.txt
h5py==3.11.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
opencv-python
scikit-image
scikit-video
av==13.1.0
open-clip-torch==2.22.0
pytorch-lightning==1.9.5
tensorboard
kornia
moviepy==1.0.3
